@@ -4,7 +4,7 @@ description: A new Eexample project.
 version: 1.0.0
 
 environment:
-  sdk: ">=2.8.1 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   magic_mirror: ^0.1.0
