@@ -1,10 +1,10 @@
 name: magic_mirror
 description: Contains annotation information and some execution content.Provides reflective actuators, and execution formatting.
-version: 0.1.0+3
+version: 0.2.0
 homepage: https://github.com/aymtools/magic_mirror
 
 environment:
-  sdk: '>=2.6.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 #dependencies:
 #  path: ^1.7.0
