@@ -4,19 +4,23 @@ version: 0.1.0+2
 homepage: https://github.com/aymtools/magic_mirror
 
 environment:
-  sdk: '>=2.6.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   magic_mirror: ^0.1.0+1
 #    path: ../magic_mirror
 
-  analyzer: ^0.41.2
-  build: ^1.6.2
-  build_resolvers: ^1.5.3
-  dart_style: ^1.3.12
+  analyzer: ^1.1.0
+  build: ^1.6.3
+  build_resolvers: ^1.5.4
+  dart_style: ^1.3.14
   source_gen: ^0.9.10
-  glob: ^1.2.0
-  path: ^1.7.0
+  glob: ^2.0.0
+  path: ^1.8.0
 
 # dev_dependencies:
 #  pedantic: ^1.9.0
+
+dependency_overrides:
+  magic_mirror:
+    path: ../magic_mirror
