@@ -1,10 +1,10 @@
-class User {
-  String name;
-  bool sex;
-  int age;
-  double height;
-
-  User.formJson(String json);
-
-  User(this.name);
-}
+// class User {
+//   String name;
+//   bool sex;
+//   int age;
+//   double height;
+//
+//   User.formJson(String json);
+//
+//   User(this.name);
+// }
