@@ -1,9 +1,10 @@
 import '../tools.dart';
 
-abstract class MMirrorException implements Exception {}
+///当前库中所有异常的父类
+abstract class MagicMirrorException implements Exception {}
 
 ///类信息无法找到异常
-class ClassNotFoundException implements MMirrorException {
+class ClassNotFoundException implements MagicMirrorException {
   ///触发异常的uri
   final String uri;
 
@@ -21,8 +22,27 @@ class ClassNotFoundException implements MMirrorException {
   }
 }
 
+///类信息无法找到异常
+class ClassNotConfigException implements MagicMirrorException {
+  ///触发异常的uri
+  final Type type;
+
+  ///其他信息
+  final message;
+
+  ClassNotConfigException(this.type, {this.message});
+
+  @override
+  String toString() {
+    final def =
+        'ClassNotConfigException:\n For ${type} can not found Class config!';
+    if (message == null) return def;
+    return '$def\n$message';
+  }
+}
+
 ///函数未找到异常
-class NoSuchFunctionException implements MMirrorException {
+class NoSuchFunctionException implements MagicMirrorException {
   ///所发生异常时的类型
   final Type type;
 
@@ -43,7 +63,7 @@ class NoSuchFunctionException implements MMirrorException {
 }
 
 ///属性未找到异常
-class NoSuchFieldException implements MMirrorException {
+class NoSuchFieldException implements MagicMirrorException {
   ///所发生异常时的类型
   final Type type;
 
@@ -64,7 +84,7 @@ class NoSuchFieldException implements MMirrorException {
 }
 
 ///参数类型异常
-class IllegalArgumentException implements MMirrorException {
+class IllegalArgumentException implements MagicMirrorException {
   ///所发生异常时的类型
   final Type type;
 
@@ -92,3 +112,27 @@ class IllegalArgumentException implements MMirrorException {
     return '$def\n$message';
   }
 }
+
+// ///参数自动转换类型时出现异常
+// class ArgumentConvertException implements MagicMirrorException {
+//   ///转换的原始类型
+//   final Type from;
+//
+//   ///转换的目标类型
+//   final Type to;
+//
+//   final source;
+//
+//   ///其他信息
+//   final message;
+//
+//   ArgumentConvertException(this.from, this.to, this.source, {this.message});
+//
+//   @override
+//   String toString() {
+//     final def =
+//         'ArgumentConvertException:\nThe source:${source} cannot from $from convert to $to';
+//     if (message == null) return def;
+//     return '$def\n$message';
+//   }
+// }
