@@ -9,23 +9,23 @@ abstract class TypeConvert<From, To> {
   Type get to => To;
 
   //转换
-  To convert(From value);
+  To? convert(From value);
 }
 
 ///类型转换器 当调用属性 函数 时类型不匹配时来完成转换动作 支持正向转换与反向转换
 abstract class TypeConvertAdapter<From, To> extends TypeConvert<From, To> {
   //反转
-  From reverse(To value);
+  From? reverse(To value);
 }
 
 ///定义自定义的转换器
 class TypeAdapter extends MClass {
-  const TypeAdapter({String adapterName})
+  const TypeAdapter({String adapterName = ''})
       : super(
-            key: adapterName == null || adapterName == ''
+            key: adapterName == ''
                 ? 'typeAdapter://mirror.aymtools.com/'
                 : 'typeAdapter://mirror.aymtools.com/' + adapterName,
-            keyGenType: adapterName == null || adapterName == ''
+            keyGenType: adapterName == ''
                 ? MClass.KEY_GEN_TYPE_BY_SEQUENCE_URI
                 : MClass.KEY_GEN_TYPE_BY_URI,
             needAssignableFrom: const <Type>[TypeConvert]);
@@ -46,13 +46,13 @@ class LoadTypeAdapter implements Initializer {
 @TypeAdapter(adapterName: 'Int2String')
 class Int2String extends TypeConvertAdapter<int, String> {
   @override
-  String convert(int value) {
+  String? convert(int value) {
     return value.toString();
   }
 
   @override
-  int reverse(String value) {
-    if (value == null || value.isEmpty) return 0;
+  int? reverse(String value) {
+    if (value.isEmpty) return null;
     return int.tryParse(value);
   }
 }
@@ -61,12 +61,12 @@ class Int2String extends TypeConvertAdapter<int, String> {
 @TypeAdapter(adapterName: 'Boolean2String')
 class Boolean2String extends TypeConvertAdapter<bool, String> {
   @override
-  String convert(bool value) {
+  String? convert(bool value) {
     return value.toString();
   }
 
   @override
-  bool reverse(String value) {
+  bool? reverse(String value) {
     return ('true' == value);
   }
 }
@@ -75,13 +75,13 @@ class Boolean2String extends TypeConvertAdapter<bool, String> {
 @TypeAdapter(adapterName: 'Double2String')
 class Double2String extends TypeConvertAdapter<double, String> {
   @override
-  String convert(double value) {
+  String? convert(double value) {
     return value.toString();
   }
 
   @override
-  double reverse(String value) {
-    if (value == null || value.isEmpty) return 0;
+  double? reverse(String value) {
+    if (value.isEmpty) return null;
     return double.tryParse(value);
   }
 }
@@ -90,12 +90,12 @@ class Double2String extends TypeConvertAdapter<double, String> {
 @TypeAdapter(adapterName: 'Int2Double')
 class Int2Double extends TypeConvertAdapter<int, double> {
   @override
-  double convert(int value) {
+  double? convert(int value) {
     return value.toDouble();
   }
 
   @override
-  int reverse(double value) {
+  int? reverse(double value) {
     return value.toInt();
   }
 }
@@ -104,12 +104,12 @@ class Int2Double extends TypeConvertAdapter<int, double> {
 @TypeAdapter(adapterName: 'Double2Bool')
 class Double2Bool extends TypeConvertAdapter<double, bool> {
   @override
-  bool convert(double value) {
+  bool? convert(double value) {
     return value == 1;
   }
 
   @override
-  double reverse(bool value) {
+  double? reverse(bool value) {
     return value ? 1 : 0;
   }
 }
@@ -118,12 +118,12 @@ class Double2Bool extends TypeConvertAdapter<double, bool> {
 @TypeAdapter(adapterName: 'Int2Bool')
 class Int2Bool extends TypeConvertAdapter<int, bool> {
   @override
-  bool convert(int value) {
+  bool? convert(int value) {
     return value == 1;
   }
 
   @override
-  int reverse(bool value) {
+  int? reverse(bool value) {
     return value ? 1 : 0;
   }
 }
